requirements.txt
torch>=2.0.0
transformers>=4.36.0
gradio==3.23.0
accelerate>=0.25.0
openai==0.28.0
python-dotenv==0.19.0
pydantic==1.10.8
starlette==0.22.0
fastapi==0.88.0
httpx==0.24.1
anyio==3.7.1 